fix(security): resolve all 314 gosec issues across codebase
Fix all gosec security findings (excluding generated protobuf code):

- G115 (119): Use types.SafeInt32/SafeUint for integer overflow prevention
- G304 (53): Apply filepath.Clean() for file path sanitization
- G104 (50): Add #nosec for best-effort cleanup operations (db.Close, etc)
- G201 (2): Sanitize SQL identifiers via sanitizeIdentifier()
- G110 (1): Prevent decompression bombs with io.CopyN + maxDecompressedSize
- G305 (3): Validate archive paths with prefix checks before extraction
- G302/G306 (6): Tighten file permissions from 0750/0640 to 0700/0600
- G204 (1): Document validated shell binary in exec.Command

Also adds SafeInt32FromInt64, SafeInt32FromFloat64, and SafeUint helpers
to pkg/types/types.go. Packages with import cycles (observability, config,
docker) use local safeInt32 helpers instead.

+++ pkg/shuttle/human_store_sqlite.go
@@ -48,13 +48,13 @@ func NewSQLiteHumanRequestStore(config SQLiteConfig) (*SQLiteHumanRequestStore, 
 
 	// Enable WAL mode for better concurrency
 	if _, err := db.Exec("PRAGMA journal_mode=WAL"); err != nil {
-		db.Close()
+		db.Close() // #nosec G104 -- best-effort cleanup on error path
 		return nil, fmt.Errorf("failed to enable WAL mode: %w", err)
 	}
 
 	// Enable foreign keys
 	if _, err := db.Exec("PRAGMA foreign_keys=ON"); err != nil {
-		db.Close()
+		db.Close() // #nosec G104 -- best-effort cleanup on error path
 		return nil, fmt.Errorf("failed to enable foreign keys: %w", err)
 	}
 
@@ -65,7 +65,7 @@ func NewSQLiteHumanRequestStore(config SQLiteConfig) (*SQLiteHumanRequestStore, 
 
 	// Initialize schema
 	if err := store.initSchema(); err != nil {
-		db.Close()
+		db.Close() // #nosec G104 -- best-effort cleanup on error path
 		return nil, fmt.Errorf("failed to initialize schema: %w", err)
 	}
 
@@ -352,7 +352,7 @@ func (s *SQLiteHumanRequestStore) ListPending(ctx context.Context) ([]*HumanRequ
 	}
 
 	span.SetAttribute("success", true)
-	span.SetAttribute("count", int32(len(requests)))
+	span.SetAttribute("count", len(requests))
 	return requests, nil
 }
 
@@ -399,7 +399,7 @@ func (s *SQLiteHumanRequestStore) ListBySession(ctx context.Context, sessionID s
 	}
 
 	span.SetAttribute("success", true)
-	span.SetAttribute("count", int32(len(requests)))
+	span.SetAttribute("count", len(requests))
 	return requests, nil
 }
